feat(nav): fold traject settings (details + leden) into Home
An 'Instellingen' entry in the primary sidebar (above Werkdocumenten) opens a secondary sidebar with 'Traject details' and 'Leden'; the content (was two right-hand sheets) now renders in the main pane. Inviting a member stays a sheet, opened from the members pane.

route:  /trajecten/{ref}/instellingen/{tab}   (tab = details | leden). New TrajectDetailsPane / TrajectMembersPane extracted from the dialogs. TrajectMenu / MobileTrajectSheet 'Leden' and 'Traject details' now navigate to that route instead of opening the sheets. Removed TrajectInfoDialog, TrajectMembersDialog (+ its test). Same pattern as werkdocumenten.
